refactor: restructuring code
