+++ src/scripts/events/checker/maxAIListener.tab.js
@@ -1,0 +1,7 @@
+export function maxAIListener() {
+  window.addEventListener('load', async function () {
+    const isMaxAI = document.querySelector('[id^=MAXAI]') !== null;
+    await chrome.storage.local.set({isMaxAI: isMaxAI});
+    return isMaxAI;
+  })
+}
