+++ src/scripts/events/uiEnhancer.tab.js
@@ -1,6 +1,8 @@
-import {appendModalContainer} from "../../units/interface/modals/appendModalContainer.tab";
-import ModalUpdate from "../../units/interface/modals/ModalUpdate.tab";
-import ModalMessage from "../../units/interface/modals/ModalMessage.tab";
+//TODO: refactor this file - dispatch the code to the right place
+
+import {appendModalContainer} from "../units/interface/modals/appendModalContainer.tab";
+import ModalUpdate from "../units/interface/modals/ModalUpdate.tab";
+import ModalMessage from "../units/interface/modals/ModalMessage.tab";
 
 
 export async function uiEnhancer(domain) {
