+++ src/scripts/process/detectPageLoad/detectPageLoad.tab.js
@@ -1,0 +1,7 @@
+import {maxAIListener} from "../../events/checker/maxAIListener.tab";
+
+export function detectPageLoad(domain) {
+  if (domain.name === "MaxAIGoogle") {
+    return maxAIListener();
+  } else return true;
+}
