+++ src/scripts/process/setupBackground/setLifecycleActions.js
@@ -1,0 +1,6 @@
+import {setInstalledAction, setUninstalledAction} from "../../events/background/lifecycleActions.bg";
+
+export function setLifecycleActions() {
+  setInstalledAction();
+  setUninstalledAction();
+}
