+++ src/scripts/units/icon/handleIconClick.bg.js
@@ -1,11 +1,11 @@
 /***
  * This is the code for the extension to run when the icon is clicked
  */
-export function clickActionListener() {
+export function handleIconClick() {
   chrome.action.onClicked.addListener(async (tab) => {
     // console.log("entered clickActionListener")
     try {
-      chrome.storage.local.set({isInjecting: true});
+      await chrome.storage.local.set({isInjecting: true});
       // console.log(chrome.storage.local.get(['isInjecting']));
 
       await chrome.scripting.executeScript({
